+++ pykechain/enums.py
@@ -178,6 +178,13 @@ class ActivityRootNames(Enum):
     APP_ROOT = 'APP_ROOT'
 
 
+activity_root_name_by_classification = {
+    ActivityClassification.WORKFLOW: ActivityRootNames.WORKFLOW_ROOT,
+    ActivityClassification.CATALOG: ActivityRootNames.CATALOG_ROOT,
+    ActivityClassification.APP: ActivityRootNames.APP_ROOT,
+}
+
+
 class ComponentXType(Enum):
     """The various inspectortypes supported in the customized task in KE-chain.
 
@@ -888,3 +895,17 @@ class ImageFitValue(Enum):
 
     CONTAIN = 'contain'
     COVER = 'cover'
+
+
+class WidgetTitleValue(Enum):
+    """
+    Options to configure the title of a widget.
+
+    :cvar DEFAULT: Use the default title of the widget type.
+    :cvar NO_TITLE: Show no title.
+    :cvar CUSTOM_TITLE: Show a custom title text.
+    """
+
+    DEFAULT = 'Default'
+    NO_TITLE = 'No title'
+    CUSTOM_TITLE = 'Custom title'