+++ pykechain/models/widgets/widgets_manager.py
@@ -144,6 +144,9 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
+        if 'parent_widget' in kwargs:
+            kwargs['parent'] = kwargs.pop('parent_widget')
+
         widget = self._client.create_widget(*args, activity=self._activity_id, **kwargs)
 
         if kwargs.get('order') is None:
@@ -233,10 +236,6 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
         # Check whether the part_model is uuid type or class `Part`
         part_model = _retrieve_object(obj=part_model, method=self._client.model)  # type: 'Part2'  # noqa
         parent_instance = _retrieve_object_id(obj=parent_instance)  # type: 'Part2'  # noqa
@@ -274,7 +273,7 @@ class WidgetsManager(Iterable):
         if all_writable and not writable_models:
             writable_models = part_model.properties
 
-        meta, title = _set_title(meta, title, default_title=part_model.name)
+        meta, title = _set_title(meta, title=title, default_title=part_model.name, **kwargs)
 
         widget = self.create_widget(
             widget_type=WidgetTypes.SUPERGRID,
@@ -387,10 +386,6 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
         # Check whether the part_model is uuid type or class `Part`
         part_model = _retrieve_object(obj=part_model, method=self._client.model)  # type: 'Part2'  # noqa
         parent_instance_id = _retrieve_object_id(obj=parent_instance)  # type: text_type
@@ -441,7 +436,7 @@ class WidgetsManager(Iterable):
         if all_writable and not writable_models:
             writable_models = part_model.properties
 
-        meta, title = _set_title(meta, title, default_title=part_model.name)
+        meta, title = _set_title(meta, title=title, default_title=part_model.name, **kwargs)
 
         widget = self.create_widget(
             widget_type=WidgetTypes.FILTEREDGRID,
@@ -483,11 +478,6 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
-
         attachment_property = _retrieve_object(attachment_property,
                                                method=self._client.property)  # type: 'Property2'
         meta = _initiate_meta(kwargs, activity=self._activity_id)
@@ -513,7 +503,8 @@ class WidgetsManager(Iterable):
                 warnings.warn('Argument "{}" is no longer supported as input to `add_attachment_viewer`.'.format(
                     deprecated_kw), Warning)
 
-        meta, title = _set_title(meta, title, default_title=attachment_property.name)
+        meta, title = _set_title(meta, title=title, default_title=attachment_property.name, **kwargs)
+
         widget = self.create_widget(
             widget_type=WidgetTypes.ATTACHMENTVIEWER,
             meta=meta,
@@ -667,11 +658,6 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
-
         # Check whether the part_model is uuid type or class `Part`
         part_instance = _retrieve_object(part_instance, method=self._client.part)  # type: 'Part2'  # noqa: F821
 
@@ -696,7 +682,7 @@ class WidgetsManager(Iterable):
             "showHeaders": show_headers,
         })
 
-        meta, title = _set_title(meta, title, default_title=part_instance.name)
+        meta, title = _set_title(meta, title=title, default_title=part_instance.name, **kwargs)
 
         if all_readable and not readable_models:
             readable_models = part_instance.model().properties
@@ -753,13 +739,7 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
-
         # Check whether the script is uuid type or class `Service`
-
         service = _retrieve_object(obj=service, method=self._client.service)  # type: 'Service'  # noqa
 
         meta = _initiate_meta(kwargs=kwargs, activity=self._activity_id)
@@ -784,7 +764,7 @@ class WidgetsManager(Iterable):
             'showLog': kwargs.get('show_log', True)
         })
 
-        meta, title = _set_title(meta, title, default_title=service.name)
+        meta, title = _set_title(meta, title=title, default_title=service.name, **kwargs)
 
         widget = self.create_widget(
             widget_type=WidgetTypes.SERVICE,
@@ -820,17 +800,12 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
-
         if not isinstance(html, (str, text_type)):
             raise IllegalArgumentError("Text injected in the HTML widget must be string. Type is: {}".
                                        format(type(html)))
 
         meta = _initiate_meta(kwargs, activity=self._activity_id)
-        meta, title = _set_title(meta, title, default_title=WidgetTypes.HTML)
+        meta, title = _set_title(meta, title=title, default_title=WidgetTypes.HTML, **kwargs)
 
         meta["htmlContent"] = html
 
@@ -869,11 +844,6 @@ class WidgetsManager(Iterable):
         :raises APIError: When the widget could not be created.
 
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
-
         from pykechain.models import Service
         if isinstance(notebook, Service):
             notebook_id = notebook.id
@@ -891,7 +861,7 @@ class WidgetsManager(Iterable):
             kwargs['custom_height'] = kwargs.pop('height')
 
         meta = _initiate_meta(kwargs=kwargs, activity=self._activity_id)
-        meta, title = _set_title(meta, title, default_title=notebook.name)
+        meta, title = _set_title(meta, title=title, default_title=notebook.name, **kwargs)
 
         meta.update({
             'serviceId': notebook_id
@@ -1062,13 +1032,8 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
-
         meta = _initiate_meta(kwargs=kwargs, activity=self._activity_id)
-        meta, title = _set_title(meta, title, default_title=WidgetTypes.MULTICOLUMN)
+        meta, title = _set_title(meta, title=title, default_title=WidgetTypes.MULTICOLUMN, **kwargs)
 
         if 'height' in kwargs:
             # TODO: Pending deprecation 3.4.0.
@@ -1157,13 +1122,8 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
-
         meta = _initiate_meta(kwargs, activity=self._activity_id)
-        meta, title = _set_title(meta, title, default_title=WidgetTypes.SCOPE)
+        meta, title = _set_title(meta, title=title, default_title=WidgetTypes.SCOPE, **kwargs)
 
         if not show_all_columns and show_columns:
             if not isinstance(show_columns, (list, tuple)) and \
@@ -1253,16 +1213,10 @@ class WidgetsManager(Iterable):
         :raises IllegalArgumentError: when incorrect arguments are provided
         :raises APIError: When the widget could not be created.
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
-
         attachment_property = _retrieve_object(attachment_property,
                                                method=self._client.property)  # type: 'Property2'  # noqa
         meta = _initiate_meta(kwargs, activity=self._activity_id)
-
-        meta, title = _set_title(meta, title, default_title=attachment_property.name)
+        meta, title = _set_title(meta, title=title, default_title=attachment_property.name, **kwargs)
 
         # Add custom button text
         if not custom_button_text:
@@ -1332,13 +1286,8 @@ class WidgetsManager(Iterable):
         :type image_fit: ImageFitValue
         :return: Card Widget
         """
-        if 'custom_title' in kwargs and not title:
-            warnings.warn('`custom_title` attribute will be deprecated in version 3.4.0, please adapt your code '
-                          'accordingly to use `title`', PendingDeprecationWarning)
-            title = kwargs.pop('custom_title')
-
         meta = _initiate_meta(kwargs, activity=self._activity_id)
-        meta, title = _set_title(meta, title, default_title=WidgetTypes.CARD)
+        meta, title = _set_title(meta, title=title, default_title=WidgetTypes.CARD, **kwargs)
 
         # Process the description
         if description is False or description is None: